memuse
Trait DynamicUsage

pub trait DynamicUsage {
    // Required methods
    fn dynamic_usage(&self) -> usize;
    fn dynamic_usage_bounds(&self) -> (usize, Option<usize>);
}

Trait for measuring the dynamic memory usage of types.

Required Methods

fn dynamic_usage(&self) -> usize

Returns a best estimate of the amount of heap-allocated memory used by this type.

For most types, this will return an exact value. However, for types that use a complex allocation strategy (such as a HashMap), memuse cannot provide an exact heap allocation value, as it does not have access to the internal details and can only infer allocations from observable properties (such as the number of elements in a collection, or constants extracted from the implementation of the type). In those cases, this method returns a “best estimate” inferred from the implemented behaviour of the type. As more crates implement this trait themselves, the estimates will become more precise.

The value returned by this method will always fall between the bounds returned by DynamicUsage::dynamic_usage_bounds:

use std::collections::HashMap;
use memuse::DynamicUsage;

let a: HashMap<u8, u64> = HashMap::with_capacity(27);
let usage = a.dynamic_usage();
let (lower, upper) = a.dynamic_usage_bounds();

assert!(lower <= usage);
if let Some(upper) = upper {
    assert!(usage <= upper);
}

fn dynamic_usage_bounds(&self) -> (usize, Option<usize>)

Returns the lower and upper bounds on the amount of heap-allocated memory used by this type.

The lower bound is always precise; a type cannot allocate fewer than zero bytes, and a collection cannot allocate fewer than the number of bytes required to store the entries it holds.

The upper bound is only present if some property of the type ensures that its allocations do not exceed the bound, and is None otherwise (to indicate an unlimited upper bound).

If the type’s allocated memory is precisely known, then the lower and upper bounds will be equal.